+++ flexmeasures/cli/tests/test_data_add.py
@@ -1,3 +1,5 @@
+import pytest
+
 from flexmeasures.cli.tests.utils import to_flags
 from flexmeasures.data.models.annotations import (
     Annotation,
@@ -6,6 +8,7 @@ from flexmeasures.data.models.annotations import (
 from flexmeasures.data.models.data_sources import DataSource
 
 
+@pytest.mark.skip_github
 def test_add_annotation(app, db, setup_roles_users):
     from flexmeasures.cli.data_add import add_annotation
 
@@ -41,6 +44,7 @@ def test_add_annotation(app, db, setup_roles_users):
     )
 
 
+@pytest.mark.skip_github
 def test_add_holidays(app, db, setup_roles_users):
     from flexmeasures.cli.data_add import add_holidays