+++ flexmeasures/data/models/time_series.py
@@ -2,7 +2,6 @@ from typing import Any, List, Dict, Optional, Union, Type, Tuple
 from datetime import datetime as datetime_type, timedelta
 import json
 
-from flask import current_app
 from sqlalchemy.ext.declarative import declared_attr
 from sqlalchemy.ext.mutable import MutableDict
 from sqlalchemy.orm import Query, Session
@@ -12,6 +11,7 @@ import timely_beliefs.utils as tb_utils
 
 from flexmeasures.auth.policy import AuthModelMixin, EVERY_LOGGED_IN_USER
 from flexmeasures.data import db
+from flexmeasures.data.models.parsing_utils import parse_source_arg
 from flexmeasures.data.queries.utils import (
     create_beliefs_query,
     get_belief_timing_criteria,
@@ -23,6 +23,10 @@ from flexmeasures.data.services.time_series import (
 )
 from flexmeasures.utils.entity_address_utils import build_entity_address
 from flexmeasures.utils.unit_utils import is_energy_unit, is_power_unit
+from flexmeasures.data.models.annotations import (
+    Annotation,
+    SensorAnnotationRelationship,
+)
 from flexmeasures.data.models.charts import chart_type_to_chart_specs
 from flexmeasures.data.models.data_sources import DataSource
 from flexmeasures.data.models.generic_assets import GenericAsset
@@ -183,6 +187,48 @@ class Sensor(db.Model, tb.SensorDBMixin, AuthModelMixin):
             most_recent_events_only=True,
             one_deterministic_belief_per_event=True,
         )
+
+    def search_annotations(
+        self,
+        annotation_starts_after: Optional[datetime_type] = None,
+        annotation_ends_before: Optional[datetime_type] = None,
+        source: Optional[
+            Union[DataSource, List[DataSource], int, List[int], str, List[str]]
+        ] = None,
+        include_asset_annotations: bool = False,
+        include_account_annotations: bool = False,
+    ):
+        parsed_sources = parse_source_arg(source)
+        query = Annotation.query.join(SensorAnnotationRelationship).filter(
+            SensorAnnotationRelationship.sensor_id == self.id,
+            SensorAnnotationRelationship.annotation_id == Annotation.id,
+        )
+        if annotation_starts_after is not None:
+            query = query.filter(
+                Annotation.start >= annotation_starts_after,
+            )
+        if annotation_ends_before is not None:
+            query = query.filter(
+                Annotation.end <= annotation_ends_before,
+            )
+        if parsed_sources:
+            query = query.filter(
+                Annotation.source.in_(parsed_sources),
+            )
+        annotations = query.all()
+        if include_asset_annotations:
+            annotations += self.generic_asset.search_annotations(
+                annotation_starts_before=annotation_starts_after,
+                annotation_ends_before=annotation_ends_before,
+                source=source,
+            )
+        if include_account_annotations:
+            annotations += self.generic_asset.owner.search_annotations(
+                annotation_starts_before=annotation_starts_after,
+                annotation_ends_before=annotation_ends_before,
+                source=source,
+            )
+        return annotations
 
     def search_beliefs(
         self,
@@ -687,38 +733,3 @@ class TimedValue(object):
             resolution=resolution,
             sum_multiple=sum_multiple,
         )
-
-
-def parse_source_arg(
-    source: Optional[
-        Union[DataSource, List[DataSource], int, List[int], str, List[str]]
-    ]
-) -> Optional[List[DataSource]]:
-    """Parse the "source" argument by looking up DataSources corresponding to any given ids or names."""
-    if source is None:
-        return source
-    if not isinstance(source, list):
-        sources = [source]
-    else:
-        sources = source
-    parsed_sources: List[DataSource] = []
-    for source in sources:
-        if isinstance(source, int):
-            parsed_source = DataSource.query.filter_by(id=source).one_or_none()
-            if parsed_source is None:
-                current_app.logger.warning(
-                    f"Beliefs searched for unknown source {source}"
-                )
-            else:
-                parsed_sources.append(parsed_source)
-        elif isinstance(source, str):
-            _parsed_sources = DataSource.query.filter_by(name=source).all()
-            if _parsed_sources is []:
-                current_app.logger.warning(
-                    f"Beliefs searched for unknown source {source}"
-                )
-            else:
-                parsed_sources.extend(_parsed_sources)
-        else:
-            parsed_sources.append(source)
-    return parsed_sources